% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lm_to_tex}
\alias{lm_to_tex}
\title{lm model to TeX}
\usage{
lm_to_tex(models, confint = T, fit = T, sample.size = T)
}
\arguments{
\item{models}{List of lm models.}

\item{confint}{Logical if \code{TRUE} adds confidence interval to the table.}

\item{fit}{Logical if \code{TRUE} adds fit information (R^2) to the table.}

\item{sample.size}{Logical if \code{TRUE} adds sample size to the table.}
}
\value{
chr TeX code for the table
}
\description{
This function takes a list of lm models and outputs TeX code to create a pretty table.
}
\details{
In this version, we start simple. The list can only contain a single model (not
much of a list). We output the TeX code to the console.

Plans for later features:
1. Take list of models and put them in the same table.
2. Output to a file
3. Compile and show pdf
}
\examples{
lm_to_tex(list(model1, model2))
lm_to_tex(list(model1), confint=F, fit=F, sample.size=F)
}

