% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lmer_to_tex}
\alias{lmer_to_tex}
\title{lmer model to TeX}
\usage{
lmer_to_tex(file, models, confint = T, fit = T, sample.size = T,
  groups = T, decimals = 2)
}
\arguments{
\item{file}{chr file name for output.}

\item{models}{List of lm models.}

\item{confint}{Logical if \code{TRUE} (default) adds confidence interval to the table.}

\item{fit}{Logical if \code{TRUE} (default) adds fit information (AIC) to the table.}

\item{sample.size}{Logical if \code{TRUE} (default) adds sample size to the table.}

\item{decimals}{int Number of decimal places (2 by default).}
}
\value{
NULL. Just outputs the table as a file.
}
\description{
This function takes a list of lme4::lmer models and outputs TeX code to create a pretty table.
}
\details{
In this version, we start simple. The list can only contain a single model (not
much of a list). We output the TeX code to the console.

Plans for later features:
1. Take list of models and put them in the same table.
2. Compile and show pdf
}
\note{
We approximate p values using lmerTest::summary
}
\examples{
## Example from lme4::lmer:
## linear mixed models - reference values from older code
model1 <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
model2 <- lmer(Reaction ~ Days + I(Days^2) + (Days | Subject), sleepstudy)

lmer_to_tex("output.tex",list(model1, model2))
lmer_to_tex("output.tex",list(model1), confint=FALSE, fit=FALSE, sample.size=FALSE)
}

