% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{TeXtable}
\alias{TeXtable}
\alias{TeXtable-package}
\title{TeXtable}
\description{
TeXtable takes one or more lm or lme4::lmer tables and creates pretty LaTeX output.
}
\details{
Although packages like xtable exist, I have found them difficult to use.
I want a simple function that takes the two most common models that I use - lm and lmer -
and outputs a TeX file that can be pasted into my manuscript (or compiled and pasted as
an image if I'm working in Word).

Later, I'll modify the package to include glm and lme4::glmer.
}

